SmartStudyBuddy/requirements.txt
langchain==0.3.14
langchain-community
langchain-openai
langchain-pinecone
langchain-core
langchain-text-splitters
pinecone-client
python-dotenv
pypdf
openai
